+++ samples/sample-fastify/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/fastify-sample
 
+## 2.1.17
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/starter-openapi@2.1.0
+
 ## 2.1.16
 
 ### Patch Changes