+++ tests/jest-fastify-demo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/jest-fastify-demo
 
+## 1.1.14
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/starter-openapi@2.1.0
+
 ## 1.1.13
 
 ### Patch Changes